pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-zero-cli"
version = "0.1.0"
description = "Terminal chat interface for Agent Zero"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.50.0",
    "python-socketio[asyncio_client]>=5.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
agentzero = "agent_zero_cli.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_zero_cli"]
